moved crossentropylabel to general utils
