+++ trailmet/algorithms/binarize/__init__.py
@@ -4,7 +4,6 @@ from .BNNBN import BNNBN
 from .ReActNet import ReActNet
 from .utils import (
     Lighting,
-    CrossEntropyLabelSmooth,
     DistributionLoss,
     adjust_learning_rate,
 )
