+++ trailmet/algorithms/binarize/utils.py
@@ -10,7 +10,6 @@ from PIL import Image
 
 __all__ = [
     "Lighting",
-    "CrossEntropyLabelSmooth",
     "DistributionLoss",
     "adjust_learning_rate",
 ]
@@ -56,22 +55,6 @@ class Lighting(object):
 
     def __repr__(self):
         return self.__class__.__name__ + "()"
-
-
-# label smooth
-class CrossEntropyLabelSmooth(nn.Module):
-    def __init__(self, num_classes, epsilon):
-        super(CrossEntropyLabelSmooth, self).__init__()
-        self.num_classes = num_classes
-        self.epsilon = epsilon
-        self.logsoftmax = nn.LogSoftmax(dim=1)
-
-    def forward(self, inputs, targets):
-        log_probs = self.logsoftmax(inputs)
-        targets = torch.zeros_like(log_probs).scatter_(1, targets.unsqueeze(1), 1)
-        targets = (1 - self.epsilon) * targets + self.epsilon / self.num_classes
-        loss = (-targets * log_probs).mean(0).sum()
-        return loss
 
 
 class DistributionLoss(loss._Loss):
